refactor: Consolidate AI suugestion logic and optimize API architecture. Extract AI logic into reusable service module(ai.service.js). Add Post/notessuggest endpoint for atomic note creation with AI suggestions
