+++ controllers/note.controller.js
@@ -1,4 +1,5 @@
 const Note = require('../models/Note');
+const aiService = require('../services/ai.service');
 
 const noteController = {};
 
@@ -152,6 +153,97 @@ noteController.deleteNote = async (req, res) => {
     return res.status(400).json({
       error: 'INVALID_ID',
       message: '올바르지 않은 노트 ID입니다.',
+    });
+  }
+};
+
+// 노트 생성 with AI suggestion
+noteController.createNoteWithSuggestion = async (req, res) => {
+  const { content, images = [] } = req.body;
+  const userId = req.userId;
+
+  try {
+    // AI 제안 생성
+    const suggestions = await aiService.generateSuggestions(content, images);
+    
+    // suggestions을 노트 데이터 형식으로 변환
+    const noteData = aiService.formatNoteFromSuggestions(suggestions);
+    
+    // AI suggestions과 함께 노트 생성
+    const note = await Note.create({
+      userId,
+      ...noteData,
+    });
+
+    return res.status(201).json({
+      message: '노트가 AI 제안과 함께 생성되었습니다.',
+      note,
+    });
+  } catch (error) {
+    console.error('Error creating note with suggestion:', {
+      message: error.message,
+      stack: error.stack,
+      timestamp: new Date().toISOString(),
+    });
+
+    // validation 에러
+    if (error.message === '내용 또는 이미지가 필요합니다') {
+      return res.status(400).json({
+        error: 'VALIDATION_ERROR',
+        message: error.message,
+      });
+    }
+
+    if (error.message.includes('이미지')) {
+      return res.status(400).json({
+        error: 'VALIDATION_ERROR',
+        message: error.message,
+      });
+    }
+
+    if (error.message === 'API 키가 설정되지 않았습니다') {
+      return res.status(500).json({
+        error: 'SERVER_ERROR',
+        message: error.message,
+      });
+    }
+
+    // AI API 에러
+    if (error.statusCode === 429) {
+      return res.status(429).json({
+        error: 'RATE_LIMIT_ERROR',
+        message: 'AI 서비스 요청 한도를 초과했습니다. 잠시 후 다시 시도해주세요.',
+      });
+    }
+
+    if (error.statusCode === 408 || error.message === 'AI_TIMEOUT') {
+      return res.status(408).json({
+        error: 'TIMEOUT_ERROR',
+        message: 'AI 처리 시간이 초과되었습니다. 내용을 줄이거나 이미지 개수를 줄여주세요.',
+      });
+    }
+
+    if (error.statusCode >= 400 && error.statusCode < 500) {
+      return res.status(400).json({
+        error: 'AI_API_ERROR',
+        message: 'AI 분석 중 오류가 발생했습니다. 입력 내용을 확인해주세요.',
+      });
+    }
+
+    // MongoDB validation에러  
+    if (error.name === 'ValidationError') {
+      return res.status(422).json({
+        error: 'VALIDATION_ERROR',
+        message: error.message,
+      });
+    }
+
+    // 서버 에러
+    return res.status(500).json({
+      error: 'SERVER_ERROR',
+      message: process.env.NODE_ENV === 'development' 
+        ? `서버 오류: ${error.message}` 
+        : '노트 생성 중 오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
     });
   }
 };
